+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/jobs/ActiveScanJob.java
@@ -648,6 +648,10 @@ public class ActiveScanJob extends AutomationJob {
         }
 
         public Integer getThreadPerHost() {
+            if (threadPerHost <= 0) {
+                // Dont return zero or less - this will cause problems
+                return null;
+            }
             return threadPerHost;
         }